Restructure Travis tests to be more transparent
Currently coverage information is submitted in the `after_success` block
of the Travis script, as it is in most packages. This makes sense for
the vast majority of packages, since if coverage submission fails you
don't want it to fail your build. However, that's exactly what we want
to happen if submission fails here.

This commit restructures the tests to put the coverage submission in the
`script` block. This will cause the build to fail if submission fails.