Deprecate verbose flag, tweak some logging code
There is no point in Coverage.LCOV.readfolder and Coverage.clean_folder
reporting every file they did *not* process; this just clogs up build logs,
and almost always is useless; so change this from @info to @debug.

Also, by default user won't want to know the JSON replies of Codecov and
Coveralls, nor the exact URL we used to submit to Codecov, so change those
from @info to @debug as well.

On the other hand, knowing which files actually are being processed *is*
useful, so add an @info message for that (I recently had an exception thrown
in that code and had a hard time figuring out which file was being proceed, so
this will help).

Finally, deprecate the `verbose` keyword flag; instead, users should adjust the
logging level to suppress or show various messages.